From:		Jon Kåre Hellan <hellan@acm.org> Subject: 	Re: GTK patch for Soundmodem configapp Date: 		Mon, 16 Mar 2009 22:41:39 +0100
ere's the final phase of the configapp cleanup. I've made it
deprecation clean as of gtk 2.16 / glib 2.20. It now requires gtk 2.6. I
think this should be OK - that version is more than 4 years old. We test
for gtk version and use the new tooltips API if available (2.12 and
newer). The conditional compiles are unsightly, but they let us build
with the .._DISABLE_DEPRECATED flags on, while not shutting out users
with older gtk.

There are a lot of pointless renamings from gtk_object_ to g_object_,
and the new widgets aren't any easier to work with with the old ones.
But sooner or later, the gtk people are going to actually drop the
deprecated widgets.

I also fixed a bug in the transition to libxml2. It would write the xml
file as a single line with no linebreaks. This has been fixed.

Joop: May be hold back on making a Debian version and give Thomas a
chance to roll a new version, if he wants to.

73
Jon LA4RT