From: Reuben Wells <row@hopgarden.net>
I thought this was the reason I was having problems getting it working
but upon further investigation, it just looks like an issue with the
configuration file XML parsing in soundcard/main.c and these errors
can be eliminated with the patch below. I've since managed to resolve
the problem I had and soundmodem is working just fine. If you could
include this change, it might save a few other people time looking
into a non-existent issue.