Hi Thomas,
I have written a patch for soundmodem to control PTT using the GPIO pin on
CM108-based USB audio dongles and compatibles.  I'm not sure what to do next.

The patch is not entirely complete because I need to test the validity of the
selected CM108 device.  This is indicated by a 'fixme' comment.  I found a
lot of useful info. on the web, especially here:
http://mgrela.rootnode.net/yaesu-usb-rif/
How do I add credit for that?  This person shows no contact details.

I think what I need to do is release the patch and ask others to test it and
comment on it, perhaps improve it.  I am not sure how to release a patch, and
in this case, where I should release it.  I think maybe the hamlib group,
especially since my patch modifies the hamlib conditional compilation (but I
think it works).  Since you are the author, and I have found no other
repository for the source, you should have the final say.  Your advice would
be appreciated.

I have had this software for a while now, and I was hoping to modify two
dongles for PTT control and test two stations with two radios passing APRS
messages back and forth.  Unfortunately I have limited time and equipment so
this is not happening as quickly as I would like.  As it stands, there may be
sufficient interest for someone else to test it much quicker than I can.

73,

Andrew Errington
ZL3AME