Hi Thomas,
My initial testing is successful!  All I did was added VID:PID for the chips I
have (not C-Media CM108, but compatibles from the 3S Company) and ran
soundmodemconfig to test PTT using the diagnostics.  Please find another
patch attached which adds the IDs for SSS1621/SSS1623.  It's a bit ugly, so
feel free to change it to your style.

Here is a list of IDs in case that is preferable to you.

; CM108/CM109/CM119
0d8c:0008
0d8c:0009
0d8c:000a
0d8c:000b
0d8c:000c
0d8c:000d
0d8c:000e
0d8c:000f

; SSS1621
0c76:1605
; SSS1623
0c76:1607
0c76:160b

As I said, I don't have two radios to try, but I will try Xastir with just one
(to verify that Xastir will assert PTT when it needs to Tx).  I will also
announce my work and your new version on linux-hams, which is where the idea
was first discussed some time ago (maybe 3 years!).  Next I will announce it
on Hamlib-developer, since the patch 'moves' their code a little bit.
Finally I will announce it on Xastir, since that is probably the most useful
client who can use it.  Actually, I recently encountered aprsg[1], which
might be able to take advantage of this too.

I hope it will work for other people.  I assume you will brace yourself for an
onslaught of bug reports and fixes.  If that happens I apologise for
upsetting the equilibrium with my crazy ideas. :)

Thank you and 73,

Andrew

[1] http://pakettiradio.net/aprsg/